Code restructure
Restructuring to make code easier to maintain - moving programmer-specific code to individual 'method' modules.
